anthropic_api_server/requirements.txt
fastapi
uvicorn[standard]
pydantic
llama-cpp-python
huggingface-hub
